[#112]fix: 인원 모집중이 아닌 여행은 삭제 불가처리

+++ src/main/java/com/example/TripAgora/session/service/SessionService.java
@@ -193,6 +193,11 @@ public class SessionService {
     @Transactional
     public void deleteSession(long userId, long sessionId) {
         Session session = findSessionAndVerifyOwner(userId, sessionId);
+
+        if (session.getStatus() != SessionStatus.RECRUITING) {
+            throw new SessionDeleteNotAllowedException();
+        }
+
         sessionRepository.delete(session);
     }
